refactor: uri 수정
