+++ src/main/java/roomescape/reservation/controller/AdminReservationApiController.java
@@ -13,22 +13,28 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.member.controller.response.MemberResponse;
+import roomescape.member.resolver.LoginMember;
 import roomescape.reservation.controller.request.ReservationCreateRequest;
 import roomescape.reservation.controller.request.ReservationRequest;
+import roomescape.reservation.controller.response.MemberReservationResponse;
 import roomescape.reservation.controller.response.ReservationResponse;
 import roomescape.reservation.service.ReservationService;
+import roomescape.reservation.service.WaitingService;
 
 @RestController
-@RequestMapping("/admin/reservations")
-public class ReservationApiController {
+@RequestMapping("/admin")
+public class AdminReservationApiController {
 
     private final ReservationService reservationService;
+    private final WaitingService waitingService;
 
-    public ReservationApiController(ReservationService reservationService) {
+    public AdminReservationApiController(ReservationService reservationService, WaitingService waitingService) {
         this.reservationService = reservationService;
+        this.waitingService = waitingService;
     }
 
-    @GetMapping
+    @GetMapping("/reservations")
     public ResponseEntity<List<ReservationResponse>> searchReservations(
             @RequestParam(required = false) Long memberId,
             @RequestParam(required = false) Long themeId,
@@ -41,7 +47,7 @@ public class ReservationApiController {
         return ResponseEntity.ok(reservationResponses);
     }
 
-    @PostMapping
+    @PostMapping("/reservations")
     public ResponseEntity<ReservationResponse> createReservation(@RequestBody @Valid ReservationCreateRequest request) {
         ReservationResponse response = reservationService.create(request.memberId(), new ReservationRequest(
                 request.date(), request.themeId(), request.timeId()
@@ -51,9 +57,16 @@ public class ReservationApiController {
                 .body(response);
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
         reservationService.deleteById(id);
         return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/waitings")
+    public ResponseEntity<List<MemberReservationResponse>> getMemberReservations(
+            @LoginMember MemberResponse memberResponse) {
+        List<MemberReservationResponse> waitings = waitingService.findWaitingsWithRankByMemberId(memberResponse.id());
+        return ResponseEntity.ok().body(waitings);
     }
 }
